requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
email-validator
cryptography
slowapi
pytest
httpx